Added a new struct + timestamps
This struct saves all variables for the "second window".
This decreases the clutter by having less global variables.
Also added the ability to save timestamps for each individual annotation (start and end). These have to be selected manually by pressing a button when the wanted timestamp is shown on screen.