Added fct for displaying windows of visible points
This function shows a Window for the point it gets called on.
This window includes the name of the Annotation and the note.
It is not movable and resizable.
It is also transparent so the background can still be seen.
Additionally it is connected to the Point it belongs to with a line.

Also added Vector: pointWindowSizes.
This one saves the size of the windows that are drawn by this function.