Removed Json Window Function
This Window is not needed, as all its functions are now possible with other windows, or directly with the ParmamSlots.

The only thing that was moved here, was the call of the function determine_points_to_be_shown, as this is now behind the ParamSlot that previously called this function.