Updated Timeline in List
Here I changed the way the timeline works.

For this I added an ImGui::PlotLines() that draws a line with 100 steps and each step is calculated if the time span of the annotation is inside of it and if it is, then the Value of that time step is set to 1, otherwise it will stay 0.

Changed the location of the Tool-tips to the first line of the table.
This also allows the usage of a "global" timeline that shows where in the animation the currently rendered frame is.
This allows the user to see if an Annotation is going to be visible soon or if it is only visible at an earlier point in time.