Changed Saving and Loading of Json
This was a bigger change.
I removed most of the old code that handled the adding of new points to the Vectors and Json_obj as well as the saving to the file.
Instead the new functions now save the new Annotations directly to the vector and then add a new entry in the json_obj.

For loading from the json file, now the old Annotations will no longer be overwritten. This means no Annotations will be lost when importing a json file.