Added Undo Button when editing Annotations
This is a sort of Undo button for all not yet saved edits of Annotations.
This works by importing the values of the json_obj that have not been changed into the current Annotation.
This only works until the button: Update the json_obj with current values has been pressed.