Fixed Heap Corruption
There was often a heap corruption that apparently is now hopefully gone with this change.

The problem was that the code that was used before was taken from a deleted function, but it was already not working there, because of other called functions that overwrote the vectors in question, so the problem never happened before it was changed.